refactor: separate context and reducer
