+++ src/components/ProductBox/ProductBox.module.scss
@@ -60,6 +60,7 @@
     .embla__container {
       display: flex;
       width: 20rem;
+      column-gap: 1px;
     }
     .embla__slide {
       position: relative;
