+++ src/pages/me/purchasehistory.js
@@ -89,6 +89,8 @@ const EachItem = ({ item, user }) => {
     setIsRating(true)
   }
 
+  const date = item.orderAt.match(/^\d{4}-\d{2}-\d{2}/g)[0]
+
   return (
     <div className={styles.eachOrder}>
       <div className={styles.eachOrder__head}>
@@ -96,7 +98,7 @@ const EachItem = ({ item, user }) => {
           <a>{item.sellerName}</a>
         </Link>
 
-        <span>seller will be</span>
+        <span>order at {date.split("-").reverse().join("-")}</span>
       </div>
       <div className={styles.eachOrder__item}>
         <div className={styles.eachOrder__image}>
